Add files via upload
In this assignment, you will create a Java program called MakingMoney.  This program will compute the worth of an investment today after a given period of time. 

Your program should read in the investment amount (present value) in CAD, the annual interest rate, and the number of years (as an integer) and display the future investment value using the following formula:


FV = Future Value.

PV = Present Value.

r = Monthly Interest Rate

n = Number of months

 

Please show the resulting amount in both CAD and in USD using the foreign exchange rate of 1 CAD = 0.747785 USD (Jan 23, 2023 18:35 UTC).

For example, if you enter the amount $1000.00, annual interest rate 3.25%, and number of years 1, the future investment value is $1032.98 CAD or $772.45 USD.

 

Here is a sample run:



Enter the investment amount in CAD, for example 120000.95: 1000.00

Enter annual interest rate %, for example 8.25: 3.25

Enter number of years as an integer, for example 5: 1

Given an investment amount of $1000.0 CAD, an annual interest rate of 3.25%, and 1 year(s) of growth,

the future value will be $1032.98 CAD or $772.45 USD.

 

Notes:

To calculate the monthly interest rate, convert the annual interest rate from percent to decimal and divide by 12.
To truncate floating point numbers to two decimal places, you can multiply the value by 100, cast it to an integer, and then divide it by 100.00.
Given the same input (in blue above), your output should match the sample run exactly.