Add files via upload
The Candy Shop


A candy shop has 3 types of candy available for purchase: gumballs, lollipops, and chocolates. Each type of candy has its own price and a discount can be applied based on the amount purchased.

Write a program named CandyShop that asks the user for the type of candy they would like to purchase and the quantity. Based on their selection and the quantity purchased, calculate and display the total price of their purchase, including any discounts that may apply.

The prices are as follows:

Gumballs: $0.25 each
Lollipops: $0.50 each
Chocolates: $1.00 each
The discounts are as follows:

Gumballs:

Buy 10 or more gumballs and get a 20% discount
Buy 20 or more gumballs and get a 25% discount
Buy 30 or more gumballs and get a 35% discount
Lollipops

Buy 5 or more lollipops and get a 10% discount
Buy 10 or more lollipops and get a 20% discount
Buy 15 or more lollipops and get a 25% discount
Chocolates

Buy 2 or more chocolates and get a 10% discount
Buy 4 or more chocolates and get a 20% discount
Buy 8 or more chocolates and get a 30% discount

Here is a sample run:



What type of candy would you like to purchase?

1. Gumballs

2. Lollipops

3. Chocolates



Selection: 2

How many lollipops would you like to purchase? 9

The total cost of your purchase is: $4.05 (discount applied: $0.45)