Add files via upload
Mad Libs is a word game where you substitute words in a story to create a silly narrative. In this assignment, you will write a program that implements a Mad Libs game. The game will prompt the user to enter a set of words, such as nouns, verbs, adjectives, etc., and then use those words to fill in the blanks in a story.